+++ engine/apps/alerts/tasks/acknowledge_reminder.py
@@ -72,10 +72,12 @@ def acknowledge_reminder_task(alert_group_pk: int, unacknowledge_process_id: str
             (alert_group.pk, unacknowledge_process_id), countdown=acknowledge_reminder_timeout
         )
 
-    log_record = alert_group.log_records.create(
-        type=AlertGroupLogRecord.TYPE_ACK_REMINDER_TRIGGERED, author=alert_group.acknowledged_by_user
-    )
-    transaction.on_commit(partial(send_alert_group_signal.delay, log_record.pk))
+    with transaction.atomic():
+        log_record = alert_group.log_records.create(
+            type=AlertGroupLogRecord.TYPE_ACK_REMINDER_TRIGGERED, author=alert_group.acknowledged_by_user
+        )
+        task_logger.info(f"created log record {log_record.pk}, sending signal...")
+        transaction.on_commit(partial(send_alert_group_signal.delay, log_record.pk))
 
 
 @shared_dedicated_queue_retry_task(autoretry_for=(Exception,), retry_backoff=True, max_retries=MAX_RETRIES)