+++ engine/apps/api/tests/test_team.py
@@ -166,7 +166,7 @@ def test_teams_number_of_users_currently_oncall_attribute_works_properly(
     team3 = make_team(organization)
 
     team1.users.set([user1, user2, user3])
-    team2.users.set([user1])
+    team2.users.set([user1, user2])
     team3.users.set([user3])
 
     def _make_schedule(team=None, oncall_users=None):
@@ -193,7 +193,9 @@ def test_teams_number_of_users_currently_oncall_attribute_works_properly(
             schedule.refresh_ical_file()
             schedule.refresh_ical_final_schedule()
 
+    # create two schedules for team 1 to make sure that every user is calculated only once per team
     _make_schedule(team=team1, oncall_users=[user1, user2])
+    _make_schedule(team=team1, oncall_users=[user1, user3])
     _make_schedule(team=team2, oncall_users=[user1])
     _make_schedule(team=team3, oncall_users=[])
 
@@ -203,7 +205,7 @@ def test_teams_number_of_users_currently_oncall_attribute_works_properly(
     response = client.get(url, format="json", **make_user_auth_headers(user1, token))
 
     number_of_oncall_users = {
-        team1.public_primary_key: 2,
+        team1.public_primary_key: 3,
         team2.public_primary_key: 1,
         team3.public_primary_key: 0,
         NO_TEAM_VALUE: 0,  # this covers the case of "No team"