+++ engine/engine/urls.py
@@ -50,16 +50,14 @@ if settings.FEATURE_PROMETHEUS_EXPORTER_ENABLED:
         path("metrics/", include("apps.metrics_exporter.urls")),
     ]
 
+if settings.FEATURE_TELEGRAM_INTEGRATION_ENABLED:
+    urlpatterns += [
+        path("telegram/", include("apps.telegram.urls")),
+    ]
+
 if settings.FEATURE_SLACK_INTEGRATION_ENABLED:
     urlpatterns += [
         path("api/internal/v1/slack/", include("apps.slack.urls")),
-    ]
-
-if settings.FEATURE_TELEGRAM_INTEGRATION_ENABLED:
-    urlpatterns += [path("telegram/", include("apps.telegram.urls"))]
-
-if settings.FEATURE_SLACK_INTEGRATION_ENABLED:
-    urlpatterns += [
         path("slack/", include("apps.slack.urls")),
     ]