+++ grafana-plugin/e2e-tests/utils/navigation.ts
@@ -2,7 +2,14 @@ import type { Page } from '@playwright/test';
 
 import { BASE_URL } from './constants';
 
-type OnCallPage = 'alert-groups' | 'integrations' | 'escalations' | 'schedules' | 'users' | 'insights';
+type OnCallPage =
+  | 'alert-groups'
+  | 'integrations'
+  | 'escalations'
+  | 'schedules'
+  | 'outgoing_webhooks'
+  | 'users'
+  | 'insights';
 
 const _goToPage = async (page: Page, url = '') => page.goto(`${BASE_URL}${url}`);
 
@@ -10,5 +17,6 @@ export const goToGrafanaPage = async (page: Page, url = '') => _goToPage(page, u
 
 export const goToOnCallPage = async (page: Page, onCallPage: OnCallPage) => {
   await _goToPage(page, `/a/grafana-oncall-app/${onCallPage}`);
+  await page.waitForLoadState('networkidle');
   await page.waitForTimeout(1000);
 };