+++ grafana-plugin/src/pages/insights/scenes/NewAlertGroupsNotificationsInTotal.tsx
@@ -10,7 +10,7 @@ export default function getNewAlertGroupsNotificationsInTotalScene({ datasource 
       {
         editorMode: 'code',
         exemplar: false,
-        expr: 'sort_desc(max_over_time(sum by(username) (avg without(pod, stack)($user_was_notified_of_alert_groups_total{slug=~"$stack"}))[1d:]))',
+        expr: 'sort_desc(max_over_time(sum by(username) (avg without(pod, instance)($user_was_notified_of_alert_groups_total{slug=~"$stack"}))[1d:]))',
         format: 'table',
         instant: true,
         legendFormat: '__auto',